# docs/swagger.yaml
definitions:
...
  go-file-server_internal_service.CopyReq:
    properties:
      destDir:
        type: string
      opId:
        description: 'Optional: client provided operation ID'
        type: string
      sources:
        items:
          type: string
        type: array
    required:
    - destDir
    - sources
    type: object
  go-file-server_internal_service.CreateFolderReq:
    properties:
      dir:
        description: Target directory path (defaults to root if empty).
        type: string
      folderName:
        type: string
      opId:
        type: string
    required:
    - folderName
    type: object
  go-file-server_internal_service.DeleteReq:
    properties:
      opId:
        type: string
      sources:
        items:
          type: string
        type: array
    required:
    - sources
    type: object
  go-file-server_internal_service.MoveReq:
    properties:
      destDir:
        type: string
      opId:
        type: string
      sources:
        items:
          type: string
        type: array
    required:
    - destDir
    - sources
    type: object
...
  go-file-server_internal_service.PropertiesReq:
    properties:
      sources:
        items:
          type: string
        type: array
    required:
    - sources
    type: object
  go-file-server_internal_service.PropertiesResponse:
    properties:
      accessedAt:
        type: string
      contains:
        $ref: '#/definitions/go-file-server_internal_service.PropertiesContains'
      createdAt:
        type: string
      location:
        type: string
      modifiedAt:
        type: string
      name:
        description: null if multiple
        type: string
      totalSizeBytes:
        type: integer
      type:
        description: '"file", "directory", or "multiple"'
        type: string
    type: object
...
host: localhost:3333
schemes:
  - http
info:
  contact:
    name: GoNet Drive Support
    url: https://github.com
  description: GoNet Drive is a self-hosted cloud file storage and sharing service.
  termsOfService: https://example.com/terms
  title: GoNet Drive API
  version: "1.0"
paths:
...
  /api/share/file/ws:
    get:
      description: Establish a WebSocket connection for real-time file operation progress
        updates within a shared context.
      parameters:
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
      security:
      - ShareAuth: []
      summary: Share WebSocket
      tags:
      - Share Files
  /api/share/file/music/play/file/{filepath}:
    get:
      description: Stream a music/audio file from a shared path. Supports HTTP range
        requests.
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - audio/*
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Serve Share Music
      tags:
      - Share Files
  /api/share/file/video/play/file/{filepath}:
    get:
      description: Stream a video file from a shared path with HTTP range support
        and optional transcoding.
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Serve Share Video
      tags:
      - Share Files
  /api/share/file/video/thumbnail/file/{filepath}:
    get:
      description: Generate and serve a thumbnail image for a video file from a shared
        path.
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Share Video Thumbnail
      tags:
      - Share Files
  /api/share/file/photo/play/file/{filepath}:
    get:
      description: Serve a photo/image file from a shared path.
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - image/*
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Serve Share Photo
      tags:
      - Share Files
  /api/share/file/photo/thumbnail/file/{filepath}:
    get:
      description: Generate and serve a thumbnail for a photo/image file from a shared
        path.
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Share Photo Thumbnail
      tags:
      - Share Files
  /api/share/file/document/read/file/{filepath}:
    get:
      description: Serve a text document file from a shared path with appropriate
        MIME type.
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Serve Share Document
      tags:
      - Share Files
  /api/share/file/copy:
    post:
      consumes:
      - application/json
      parameters:
      - description: Copy request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.CopyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Copy Files in Share
      tags:
      - Share Files
  /api/share/file/create-folder:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create folder request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.CreateFolderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Create Folder in Share
      tags:
      - Share Files
  /api/share/file/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Delete request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.DeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Soft Delete Share Files
      tags:
      - Share Files
  /api/share/file/delete-permanent:
    post:
      consumes:
      - application/json
      parameters:
      - description: Delete request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.DeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Permanent Delete Share Files
      tags:
      - Share Files
  /api/share/file/download:
    get:
      description: Download one or more files/directories from a shared path. Single
        files are served directly; multiple items are packaged as ZIP.
      parameters:
      - collectionFormat: csv
        description: Source file paths
        in: query
        items:
          type: string
        name: source
        required: true
        type: array
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - application/octet-stream
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Download Share Files
      tags:
      - Share Files
  /api/share/file/list:
    get:
      description: List files and directories within the authorized share path.
      parameters:
      - description: Directory path
        in: query
        name: path
        type: string
      - description: 'Sort by: name, size, modified'
        in: query
        name: sort
        type: string
      - description: 'Sort order: asc, desc'
        in: query
        name: order
        type: string
      - description: 'Show hidden files: true, false'
        in: query
        name: showHidden
        type: string
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: List Share Files
      tags:
      - Share Files
  /api/share/file/move:
    post:
      consumes:
      - application/json
      parameters:
      - description: Move request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.MoveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Move Files in Share
      tags:
      - Share Files
  /api/share/file/properties:
    post:
      consumes:
      - application/json
      description: Get metadata for files/directories in a shared path.
      parameters:
      - description: Share ID
        in: header
        name: X-Share-Id
        type: string
      - description: Share ID (alternative)
        in: query
        name: share_id
        type: string
      - description: Properties request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.PropertiesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-file-server_internal_service.PropertiesResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Share File Properties
      tags:
      - Share Files
  /api/share/file/rename:
    post:
      consumes:
      - application/json
      parameters:
      - description: Rename request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-file-server_internal_service.RenameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ShareAuth: []
      summary: Rename File in Share
      tags:
      - Share Files
...
  /api/user/document/read/file/{filepath}:
    get:
      description: Serve a text document file with appropriate MIME type.
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Serve Document
      tags:
      - Media
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: JWT access token stored as a cookie.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"